+++ bin/build.sh
@@ -1,9 +1,0 @@
-#!/usr/bin/env sh
-
-function build() {
-  npm install
-  composer install --no-dev --no-scripts
-  npm run build
-}
-
-build