+++ skin/assets/application-admin.js
@@ -1,2 +1,0 @@
-// Load Styles
-import './styles/application-admin.scss';