+++ skin/assets/fonts/index.js
@@ -1,5 +1,0 @@
-// import './theme_font.eot';
-// import './theme_font.svg';
-// import './theme_font.ttf';
-// import './theme_font.woff';
-// import './theme_font.ttf';