+++ template-parts/tracking/head.php
@@ -1,9 +1,0 @@
-<?php
-/**
- * Tracking codes in Head
- *
- * @package Inf_Theme\Template_Parts\Tracking
- */
-
-get_template_part( 'template-parts/tracking/parts/facebook' );
-get_template_part( 'template-parts/tracking/parts/google-tag-manager' );